Rename training2 -> training.
