+++ src/invoke_training/scripts/invoke_finetune_lora_sdxl.py
@@ -4,7 +4,7 @@ from pathlib import Path
 import yaml
 
 from invoke_training.config.pipelines.finetune_lora_config import FinetuneLoRASDXLConfig
-from invoke_training.training2.pipelines.stable_diffusion_xl.finetune_lora_sdxl import run_training
+from invoke_training.training.pipelines.stable_diffusion_xl.finetune_lora_sdxl import run_training
 
 
 def parse_args():
