+++ src/invoke_training/training/shared/checkpoints/serialization.py
