+++ src/invoke_training/training/shared/data/data_loaders/dreambooth_sdxl_dataloader.py
@@ -3,21 +3,21 @@ import typing
 from torch.utils.data import ConcatDataset, DataLoader
 
 from invoke_training.config.shared.data.data_config import DreamBoothDataLoaderConfig
-from invoke_training.training2.shared.data.data_loaders.dreambooth_sd_dataloader import (
+from invoke_training.training.shared.data.data_loaders.dreambooth_sd_dataloader import (
     InterleavedSampler,
     SequentialRangeSampler,
     ShuffledRangeSampler,
 )
-from invoke_training.training2.shared.data.data_loaders.image_caption_sdxl_dataloader import (
+from invoke_training.training.shared.data.data_loaders.image_caption_sdxl_dataloader import (
     sdxl_image_caption_collate_fn,
 )
-from invoke_training.training2.shared.data.datasets.image_dir_dataset import ImageDirDataset
-from invoke_training.training2.shared.data.datasets.transform_dataset import TransformDataset
-from invoke_training.training2.shared.data.transforms.constant_field_transform import ConstantFieldTransform
-from invoke_training.training2.shared.data.transforms.drop_field_transform import DropFieldTransform
-from invoke_training.training2.shared.data.transforms.load_cache_transform import LoadCacheTransform
-from invoke_training.training2.shared.data.transforms.sdxl_image_transform import SDXLImageTransform
-from invoke_training.training2.shared.data.transforms.tensor_disk_cache import TensorDiskCache
+from invoke_training.training.shared.data.datasets.image_dir_dataset import ImageDirDataset
+from invoke_training.training.shared.data.datasets.transform_dataset import TransformDataset
+from invoke_training.training.shared.data.transforms.constant_field_transform import ConstantFieldTransform
+from invoke_training.training.shared.data.transforms.drop_field_transform import DropFieldTransform
+from invoke_training.training.shared.data.transforms.load_cache_transform import LoadCacheTransform
+from invoke_training.training.shared.data.transforms.sdxl_image_transform import SDXLImageTransform
+from invoke_training.training.shared.data.transforms.tensor_disk_cache import TensorDiskCache
 
 
 def build_dreambooth_sdxl_dataloader(
