+++ src/invoke_training/training/shared/data/data_loaders/textual_inversion_sd_dataloader.py
@@ -3,14 +3,14 @@ import typing
 from torch.utils.data import DataLoader
 
 from invoke_training.config.shared.data.data_config import TextualInversionDataLoaderConfig
-from invoke_training.training2.shared.data.datasets.image_dir_dataset import ImageDirDataset
-from invoke_training.training2.shared.data.datasets.transform_dataset import TransformDataset
-from invoke_training.training2.shared.data.transforms.drop_field_transform import DropFieldTransform
-from invoke_training.training2.shared.data.transforms.load_cache_transform import LoadCacheTransform
-from invoke_training.training2.shared.data.transforms.sd_image_transform import SDImageTransform
-from invoke_training.training2.shared.data.transforms.shuffle_caption_transform import ShuffleCaptionTransform
-from invoke_training.training2.shared.data.transforms.tensor_disk_cache import TensorDiskCache
-from invoke_training.training2.shared.data.transforms.textual_inversion_caption_transform import (
+from invoke_training.training.shared.data.datasets.image_dir_dataset import ImageDirDataset
+from invoke_training.training.shared.data.datasets.transform_dataset import TransformDataset
+from invoke_training.training.shared.data.transforms.drop_field_transform import DropFieldTransform
+from invoke_training.training.shared.data.transforms.load_cache_transform import LoadCacheTransform
+from invoke_training.training.shared.data.transforms.sd_image_transform import SDImageTransform
+from invoke_training.training.shared.data.transforms.shuffle_caption_transform import ShuffleCaptionTransform
+from invoke_training.training.shared.data.transforms.tensor_disk_cache import TensorDiskCache
+from invoke_training.training.shared.data.transforms.textual_inversion_caption_transform import (
     TextualInversionCaptionTransform,
 )
