+++ src/invoke_training/training/shared/data/datasets/hf_dir_image_caption_dataset.py
