+++ src/invoke_training/training/shared/data/datasets/hf_hub_image_caption_dataset.py
