+++ src/invoke_training/training/shared/data/datasets/image_dir_dataset.py
