+++ src/invoke_training/training/shared/data/transforms/sd_image_transform.py
