+++ src/invoke_training/training/shared/data/transforms/sdxl_image_transform.py
