+++ src/invoke_training/training/shared/data/transforms/textual_inversion_caption_transform.py
