+++ src/invoke_training/training/shared/stable_diffusion/base_model_version.py
