+++ src/invoke_training/training/shared/stable_diffusion/model_loading_utils.py
