+++ src/invoke_training/training/shared/stable_diffusion/tokenize_captions.py
