+++ tests/invoke_training/training/shared/data/data_loaders/test_dreambooth_sd_dataloader.py
@@ -1,7 +1,7 @@
 import torch
 
 from invoke_training.config.shared.data.data_config import DreamBoothDataLoaderConfig
-from invoke_training.training2.shared.data.data_loaders.dreambooth_sd_dataloader import build_dreambooth_sd_dataloader
+from invoke_training.training.shared.data.data_loaders.dreambooth_sd_dataloader import build_dreambooth_sd_dataloader
 
 from ..image_dir_fixture import image_dir  # noqa: F401
