+++ tests/invoke_training/training/shared/data/data_loaders/test_dreambooth_sdxl_dataloader.py
@@ -1,7 +1,7 @@
 import torch
 
 from invoke_training.config.shared.data.data_config import DreamBoothDataLoaderConfig
-from invoke_training.training2.shared.data.data_loaders.dreambooth_sdxl_dataloader import (
+from invoke_training.training.shared.data.data_loaders.dreambooth_sdxl_dataloader import (
     build_dreambooth_sdxl_dataloader,
 )
