+++ tests/invoke_training/training/shared/data/data_loaders/test_image_caption_sd_dataloader.py
@@ -4,7 +4,7 @@ import torch
 
 from invoke_training.config.pipelines.finetune_lora_config import ImageCaptionDataLoaderConfig
 from invoke_training.config.shared.data.data_config import ImageTransformConfig
-from invoke_training.training2.shared.data.data_loaders.image_caption_sd_dataloader import (
+from invoke_training.training.shared.data.data_loaders.image_caption_sd_dataloader import (
     build_image_caption_sd_dataloader,
 )
