+++ tests/invoke_training/training/shared/data/data_loaders/test_textual_inversion_sd_dataloader.py
@@ -1,7 +1,7 @@
 import torch
 
 from invoke_training.config.shared.data.data_config import ImageTransformConfig, TextualInversionDataLoaderConfig
-from invoke_training.training2.shared.data.data_loaders.textual_inversion_sd_dataloader import (
+from invoke_training.training.shared.data.data_loaders.textual_inversion_sd_dataloader import (
     build_textual_inversion_sd_dataloader,
 )
