+++ tests/invoke_training/training/shared/data/transforms/test_sdxl_image_transform.py
@@ -4,7 +4,7 @@ import numpy as np
 import torch
 from PIL import Image
 
-from invoke_training.training2.shared.data.transforms.sdxl_image_transform import SDXLImageTransform
+from invoke_training.training.shared.data.transforms.sdxl_image_transform import SDXLImageTransform
 
 
 def denormalize_image(img: np.ndarray) -> np.ndarray:
