+++ tests/invoke_training/training/shared/data/transforms/test_shuffle_caption_transform.py
@@ -1,4 +1,4 @@
-from invoke_training.training2.shared.data.transforms.shuffle_caption_transform import ShuffleCaptionTransform
+from invoke_training.training.shared.data.transforms.shuffle_caption_transform import ShuffleCaptionTransform
 
 
 def test_shuffle_caption_transform():
